+++ commitlint.config.js
@@ -1,6 +1,6 @@
-module.exports = {
+export default {
   extends: ['@commitlint/config-conventional'],
   rules: {
-    'body-max-line-length': [1, 'always', 200]
-  }
+    'body-max-line-length': [1, 'always', 200],
+  },
 }