+++ src/Dataset.vue
@@ -13,37 +13,38 @@
       dsPages,
       search,
       showEntries,
-      setActive
+      setActive,
     }"
   ></slot>
 </template>
 
 <script>
-import i18n from './i18n/en.js'
 import { computed, nextTick, provide, ref, watch } from 'vue'
 import { createPagingRange, fieldFilter, fieldSorter, findAny, isEmptyObject } from './helpers'
+import i18n from './i18n/en.js'
 
 export default {
+  name: 'Dataset',
   props: {
     dsData: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     dsFilterFields: {
       type: Object,
-      default: () => ({})
+      default: () => ({}),
     },
     dsSortby: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     dsSearchIn: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     dsSearchAs: {
       type: Object,
-      default: () => ({})
+      default: () => ({}),
     },
     dsSortAs: {
       type: Object,
@@ -162,13 +163,13 @@ export default {
 
         emit(
           'update:dsData',
-          result.map((i) => props.dsData[i])
+          result.map((i) => props.dsData[i]),
         )
       },
       {
         immediate: true,
-        deep: true
-      }
+        deep: true,
+      },
     )
 
     provide('dsIndexes', dsIndexes)
@@ -178,7 +179,7 @@ export default {
     provide('datasetI18n', datasetI18n)
     provide(
       'dsData',
-      computed(() => props.dsData)
+      computed(() => props.dsData),
     )
     provide('dsRows', dsRows)
     provide('dsPages', dsPages)
@@ -200,8 +201,8 @@ export default {
       dsPages,
       search,
       showEntries,
-      setActive
+      setActive,
     }
-  }
+  },
 }
 </script>