+++ src/DatasetSearch.vue
@@ -10,7 +10,9 @@
 
 <script setup>
 import { debounce } from './helpers'
+
 import { inject, ref } from 'vue'
+import { debounce } from './helpers'
 
 const props = defineProps({
   dsSearchPlaceholder: {
@@ -30,4 +32,10 @@ const dsSearch = ref('')
 const input = debounce((value) => {
   search(value)
 }, props.wait)
+    return {
+      dsSearch,
+      input,
+    }
+  },
+}
 </script>