+++ vite.config.js
@@ -1,10 +1,9 @@
 import vue from '@vitejs/plugin-vue'
+import { resolve } from 'path'
+import { defineConfig } from 'vite'
 import { createHtmlPlugin } from 'vite-plugin-html'
 
-const path = require('path')
-const { defineConfig } = require('vite')
-
-module.exports = defineConfig({
+export default defineConfig({
   publicDir: 'public-vite',
   plugins: [
     vue(),
@@ -13,42 +12,53 @@ module.exports = defineConfig({
       inject: {
         data: {
           title: 'vue-dataset playground',
-          description: 'Playground for vue-dataset Vue.js 3'
-        }
-      }
-    })
+          description: 'Playground for vue-dataset Vue.js 3',
+        },
+      },
+    }),
   ],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, './src'),
-      '@playground': path.resolve(__dirname, './playground'),
-      '@root': path.resolve(__dirname, './'),
-      '~bootstrap': 'bootstrap'
-    }
+      '@': resolve(__dirname, './src'),
+      '@playground': resolve(__dirname, './playground'),
+      '@root': resolve(__dirname, './'),
+      '~bootstrap': 'bootstrap',
+    },
   },
   rollupInputOptions: {
-    input: path.resolve(__dirname, '/playground/main.js') // custom main
+    input: resolve(__dirname, '/playground/main.js'), // custom main
   },
   css: {
     preprocessorOptions: {
       scss: {
-        additionalData: `@import "./playground/scss/variables";`
-      }
-    }
+        additionalData: `@import "./playground/scss/variables";`,
+      },
+    },
   },
   build: {
     lib: {
-      entry: path.resolve(__dirname, 'src/index.js'),
+      entry: resolve(__dirname, 'src/index.js'),
       name: 'VueDataset',
-      fileName: (format) => `vue-dataset.${format}.js`
+      fileName: (format) => `vue-dataset.${format}.js`,
     },
     rollupOptions: {
       external: ['vue'],
       output: {
         globals: {
-          vue: 'Vue'
-        }
-      }
-    }
-  }
+          vue: 'Vue',
+        },
+      },
+    },
+  },
+  test: {
+    globals: true,
+    // Clear the mocks call count before each test so that we don't have to call vi.clearAllMocks manually - https://vitest.dev/config/#clearmocks
+    clearMocks: true,
+    environment: 'jsdom',
+    reporters: ['default'],
+    coverage: {
+      reporter: ['text', ['json-summary', { file: 'coverage-summary.json' }]],
+      include: ['src/**'],
+    },
+  },
 })